(#1224) Rename tests
